♻️ refactor: transform to OOP
