+++ utils/utils.py
@@ -1,0 +1,25 @@
+import os
+
+
+article_type_dict = {
+    0: "thoi-su",
+    1: "du-lich",
+    2: "the-gioi",
+    3: "kinh-doanh",
+    4: "khoa-hoc",
+    5: "giai-tri",
+    6: "the-thao",
+    7: "phap-luat",
+    8: "giao-duc",
+    9: "suc-khoe",
+    10: "doi-song"
+}              
+
+def create_dir(path):
+    if not os.path.exists(path):
+        os.makedirs(path)
+
+def read_file(path):
+    with open(path, encoding="utf-8") as file:
+        for line in file:
+            yield line
